demo-session1/requirements.txt
mysql-connector-python
pandas
python-dotenv
pathlib
pathlib
PyYAML
pylint
black
pytest
boto3
fsspec
s3fs
gspread 
oauth2client
googleapiclient
